Feature/setup batch service

+++ batch-service/src/main/java/org/com/batchservice/dto/response/DocumentData.java
@@ -12,5 +12,5 @@ public class DocumentData {
     private String title;
     private String category;
     private String content;
-    private float popularity;
+    private double popularity;
 }
